+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.0.5
+- a few new debugging logging messages
+- improved way to set logging level for ``slkspec``
+
 ## v0.0.4
 - incremented dependency on ``pyslk`` to ``2.3.0`` => improved workflows + new required functions
 - internal ``list`` of files needed to be retrieved was changed to type ``set`` => remove duplicate files